refactor(convert): move the implementation out of cmd into internal/convert
cmd/convert.go is cobra wiring only — flags, validation, delegation —
matching how every other command hands off to internal packages. The
walk/migration/encode logic and its behavioral tests live in
internal/convert; the command test keeps only flag validation.
